Rename append_measure_* to append_column_to_measure_grid_*
