+++ src/actions/append_column_to_measure_grid_action.cpp
@@ -2,7 +2,7 @@
 
 
 
-#include <fullscore/actions/append_measure_action.h>
+#include <fullscore/actions/append_column_to_measure_grid_action.h>
 
 #include <fullscore/models/measure_grid.h>
 
@@ -10,7 +10,7 @@
 
 
 Action::AppendMeasure::AppendMeasure(MeasureGrid *measure_grid)
-   : Base("append_measure")
+   : Base("append_column_to_measure_grid")
    , measure_grid(measure_grid)
 {}
